pyproject.toml
[project]
name = "openscholar"
version = "0.1.0"
description = "OpenScholar: Synthesizing Scientific Literature with Retrieval-Augmented LMs"
authors = [{name = "OpenScholar Team"}]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=2.2.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "peft>=0.6.0",
    "bitsandbytes>=0.41.0",
    "wandb>=0.15.0",
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "streamlit>=1.28.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.0",
    "aiohttp>=3.9.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "tiktoken>=0.5.0",
    "pymongo>=4.5.0",
    "elasticsearch>=8.10.0",
    "redis>=5.0.0",
    "minio>=7.2.0",
    "pyarrow>=14.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "rich>=13.6.0",
    "typer>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "ipython>=8.15.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "pre-commit>=3.4.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"